Refactor timestamp to shared util
