+++ time_utils.hpp
@@ -1,0 +1,11 @@
+#ifndef TIME_UTILS_HPP
+#define TIME_UTILS_HPP
+
+#include <string>
+
+/**
+ * @brief Get the current local time formatted as YYYY-MM-DD HH:MM:SS.
+ */
+std::string timestamp();
+
+#endif // TIME_UTILS_HPP
